MC-34298: Introduce separate POST and PUT API for carts/mine/items
