+++ app/code/Magento/Quote/Api/UpdateCartItemInterface.php
@@ -7,6 +7,8 @@ declare(strict_types=1);
 
 namespace Magento\Quote\Api;
 
+use Magento\Quote\Api\Data\CartItemInterface;
+
 /**
  * Interface AddCartItemInterface
  * @api
@@ -16,11 +18,11 @@ interface UpdateCartItemInterface
     /**
      * Update the specified cart item.
      *
-     * @param \Magento\Quote\Api\Data\CartItemInterface $cartItem The item.
-     * @return \Magento\Quote\Api\Data\CartItemInterface Item.
+     * @param CartItemInterface $cartItem The item.
+     * @return CartItemInterface Item.
      * @throws \Magento\Framework\Exception\NoSuchEntityException The specified cart does not exist.
      * @throws \Magento\Framework\Exception\CouldNotSaveException The specified item could not be saved to the cart.
      * @throws \Magento\Framework\Exception\InputException The specified item or cart is not valid.
      */
-    public function execute(\Magento\Quote\Api\Data\CartItemInterface $cartItem);
+    public function execute(CartItemInterface $cartItem): CartItemInterface;
 }
