+++ dev/tests/api-functional/testsuite/Magento/Quote/Api/CartItemAddTest.php
@@ -1,0 +1,76 @@
+<?php
+/**
+ *
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\Quote\Api;
+
+use Magento\Catalog\Model\Product;
+use Magento\Framework\Webapi\Rest\Request;
+use Magento\Quote\Model\Quote;
+use Magento\TestFramework\Helper\Bootstrap;
+use Magento\TestFramework\ObjectManager;
+use Magento\TestFramework\TestCase\WebapiAbstract;
+
+/**
+ * API test for add cart item
+ */
+class CartItemAddTest extends WebapiAbstract
+{
+    const SERVICE_VERSION = 'V1';
+    const SERVICE_NAME = 'quoteAddCartItemV1';
+    const RESOURCE_PATH = '/V1/carts/';
+
+    /**
+     * @var ObjectManager
+     */
+    protected $objectManager;
+
+    /**
+     * @inheritDoc
+     */
+    protected function setUp(): void
+    {
+        $this->objectManager = Bootstrap::getObjectManager();
+    }
+
+    /**
+     * @magentoApiDataFixture Magento/Checkout/_files/quote_with_address_saved.php
+     * @magentoApiDataFixture Magento/Catalog/_files/product_without_options.php
+     */
+    public function testAddItem(): void
+    {
+        /** @var  Product $product */
+        $product = $this->objectManager->create(Product::class)->load(2);
+        $productSku = $product->getSku();
+        /** @var Quote  $quote */
+        $quote = $this->objectManager->create(Quote::class);
+        $quote->load('test_order_1', 'reserved_order_id');
+        $cartId = $quote->getId();
+        $serviceInfo = [
+            'rest' => [
+                'resourcePath' => self::RESOURCE_PATH .  $cartId . '/items',
+                'httpMethod' => Request::HTTP_METHOD_POST,
+            ],
+            'soap' => [
+                'service' => self::SERVICE_NAME,
+                'serviceVersion' => self::SERVICE_VERSION,
+                'operation' => self::SERVICE_NAME . 'Execute',
+            ],
+        ];
+
+        $requestData = [
+            "cartItem" => [
+                "sku" => $productSku,
+                "qty" => 7,
+                "quote_id" => $cartId,
+            ],
+        ];
+        $this->_webApiCall($serviceInfo, $requestData);
+        $this->assertTrue($quote->hasProductId(2));
+        $this->assertEquals(7, $quote->getItemByProduct($product)->getQty());
+    }
+}
