+++ dev/tests/api-functional/testsuite/Magento/Quote/Api/CartItemRepositoryTest.php
@@ -4,13 +4,22 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
+declare(strict_types=1);
+
 namespace Magento\Quote\Api;
 
 use Magento\Catalog\Model\CustomOptions\CustomOptionProcessor;
+use Magento\Catalog\Model\Product;
 use Magento\Framework\Webapi\Rest\Request;
+use Magento\Quote\Api\Data\CartItemInterface;
 use Magento\Quote\Model\Quote;
+use Magento\TestFramework\Helper\Bootstrap;
+use Magento\TestFramework\ObjectManager;
 use Magento\TestFramework\TestCase\WebapiAbstract;
 
+/**
+ * API test for cart item repository
+ */
 class CartItemRepositoryTest extends WebapiAbstract
 {
     const SERVICE_VERSION = 'V1';
@@ -18,13 +27,16 @@ class CartItemRepositoryTest extends WebapiAbstract
     const RESOURCE_PATH = '/V1/carts/';
 
     /**
-     * @var \Magento\TestFramework\ObjectManager
+     * @var ObjectManager
      */
     protected $objectManager;
 
+    /**
+     * @inheritDoc
+     */
     protected function setUp(): void
     {
-        $this->objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
+        $this->objectManager = Bootstrap::getObjectManager();
     }
 
     /**
@@ -39,7 +51,7 @@ class CartItemRepositoryTest extends WebapiAbstract
         $output = [];
         $customOptionProcessor = $this->objectManager->get(CustomOptionProcessor::class);
 
-        /** @var  \Magento\Quote\Api\Data\CartItemInterface $item */
+        /** @var  CartItemInterface $item */
         foreach ($quote->getAllItems() as $item) {
             $customOptionProcessor->processOptions($item);
             $data = [
@@ -78,43 +90,6 @@ class CartItemRepositoryTest extends WebapiAbstract
     }
 
     /**
-     * @magentoApiDataFixture Magento/Checkout/_files/quote_with_address_saved.php
-     * @magentoApiDataFixture Magento/Catalog/_files/product_without_options.php
-     */
-    public function testAddItem()
-    {
-        /** @var  \Magento\Catalog\Model\Product $product */
-        $product = $this->objectManager->create(\Magento\Catalog\Model\Product::class)->load(2);
-        $productSku = $product->getSku();
-        /** @var Quote  $quote */
-        $quote = $this->objectManager->create(Quote::class);
-        $quote->load('test_order_1', 'reserved_order_id');
-        $cartId = $quote->getId();
-        $serviceInfo = [
-            'rest' => [
-                'resourcePath' => self::RESOURCE_PATH .  $cartId . '/items',
-                'httpMethod' => Request::HTTP_METHOD_POST,
-            ],
-            'soap' => [
-                'service' => self::SERVICE_NAME,
-                'serviceVersion' => self::SERVICE_VERSION,
-                'operation' => self::SERVICE_NAME . 'Save',
-            ],
-        ];
-
-        $requestData = [
-            "cartItem" => [
-                "sku" => $productSku,
-                "qty" => 7,
-                "quote_id" => $cartId,
-            ],
-        ];
-        $this->_webApiCall($serviceInfo, $requestData);
-        $this->assertTrue($quote->hasProductId(2));
-        $this->assertEquals(7, $quote->getItemByProduct($product)->getQty());
-    }
-
-    /**
      * @magentoApiDataFixture Magento/Checkout/_files/quote_with_items_saved.php
      */
     public function testRemoveItem()
@@ -123,7 +98,7 @@ class CartItemRepositoryTest extends WebapiAbstract
         $quote = $this->objectManager->create(Quote::class);
         $quote->load('test_order_item_with_items', 'reserved_order_id');
         $cartId = $quote->getId();
-        $product = $this->objectManager->create(\Magento\Catalog\Model\Product::class);
+        $product = $this->objectManager->create(Product::class);
         $productId = $product->getIdBySku('simple_one');
         $product->load($productId);
         $itemId = $quote->getItemByProduct($product)->getId();
@@ -147,55 +122,5 @@ class CartItemRepositoryTest extends WebapiAbstract
         $quote = $this->objectManager->create(Quote::class);
         $quote->load('test_order_item_with_items', 'reserved_order_id');
         $this->assertFalse($quote->hasProductId($productId));
-    }
-
-    /**
-     * @magentoApiDataFixture Magento/Checkout/_files/quote_with_items_saved.php
-     */
-    public function testUpdateItem()
-    {
-        /** @var Quote  $quote */
-        $quote = $this->objectManager->create(Quote::class);
-        $quote->load('test_order_item_with_items', 'reserved_order_id');
-        $cartId = $quote->getId();
-        $product = $this->objectManager->create(\Magento\Catalog\Model\Product::class);
-        $productId = $product->getIdBySku('simple_one');
-        $product->load($productId);
-        $itemId = $quote->getItemByProduct($product)->getId();
-        $serviceInfo = [
-            'rest' => [
-                'resourcePath' => self::RESOURCE_PATH .  $cartId . '/items/' . $itemId,
-                'httpMethod' => Request::HTTP_METHOD_PUT,
-            ],
-            'soap' => [
-                'service' => self::SERVICE_NAME,
-                'serviceVersion' => self::SERVICE_VERSION,
-                'operation' => self::SERVICE_NAME . 'Save',
-            ],
-        ];
-
-        if (TESTS_WEB_API_ADAPTER == self::ADAPTER_SOAP) {
-            $requestData = [
-                "cartItem" => [
-                    "qty" => 5,
-                    "quote_id" => $cartId,
-                    "itemId" => $itemId,
-                ],
-            ];
-        } else {
-            $requestData = [
-                "cartItem" => [
-                    "qty" => 5,
-                    "quote_id" => $cartId,
-                ],
-            ];
-        }
-        $this->_webApiCall($serviceInfo, $requestData);
-        $quote = $this->objectManager->create(Quote::class);
-        $quote->load('test_order_item_with_items', 'reserved_order_id');
-        $this->assertTrue($quote->hasProductId(1));
-        $item = $quote->getItemByProduct($product);
-        $this->assertEquals(5, $item->getQty());
-        $this->assertEquals($itemId, $item->getItemId());
     }
 }
